Migrating to vue3 also other plugins as axios - WIP
