+++ app/javascript/users/init.js
@@ -1,33 +1,39 @@
-import Vue from 'vue'
+// import Vue from 'vue'
+// import './plugins'
 
-import App from './components/index'
-import store from './store/store'
-import router from './router'
-
-import moment from 'moment'
-
-import VueI18n from 'vue-i18n'
-import Vuelidate from 'vuelidate'
-
-import vuetify from './plugins/vuetify'
-import './plugins'
 
 // import * as VueGoogleMaps from 'vue2-google-maps'
 import ActionCableVue from 'actioncable-vue'
-
-Vue.use(VueI18n)
-Vue.use(Vuelidate)
-
-import axios from 'axios'
+import axios from "axios"
 
 const init = function() {
+  import { createApp } from 'vue'
+  const app = createApp({})
+
+  import App from './components/index'
+  import store from './store/store'
+  import router from './router'
+
+  import moment from 'moment'
+
+  import VueI18n from 'vue-i18n'
+  import Vuelidate from 'vuelidate'
+
+  import vuetify from './plugins/vuetify'
+
+  import LoadHomepageComponents from './plugins/load_homepage_components'
+  LoadHomepageComponents(app)
+
+  app.use(VueI18n)
+  app.use(Vuelidate)
+
   let el = document.getElementById('users-vue-wrapper')
   if (el !== null) {
     let id = '#' + el.getAttribute('id')
     let props = JSON.parse(el.getAttribute('data-props'))
 
     if (props?.options?.websocketUrl) {
-      Vue.use(ActionCableVue, {
+      app.use(ActionCableVue, {
         debug: true,
         debugLevel: 'error',
         connectionUrl: props.options.websocketUrl, // or function which returns a string with your JWT appended to your server URL as a query parameter
