+++ app/javascript/users/plugins/index.js
@@ -1,5 +1,5 @@
 // Automatically included in './src/main.js'
 
-import './base'
+// import './base'
 // import './meta'
 // import './webfontloader'
