+++ package.json
@@ -30,12 +30,11 @@
     "trim-newlines": "3.0.1",
     "turbolinks": "^5.2.0",
     "uuid": "^3.4.0",
-    "vue": "^2.7.14",
+    "vue": "^3.2.47",
     "vue-i18n": "^8.28.2",
-    "vue-loader": "^15.10.1",
-    "vue-router": "^3.6.5",
-    "vue-stripe-elements-plus": "^0.3.2",
-    "vue-template-compiler": "^2.7.14",
+    "vue-loader": "^17.0.1",
+    "vue-router": "^4.1.6",
+    "vue-stripe-elements-plus": "^1.0.3",
     "vuelidate": "^0.7.7",
     "vuetify": "^2.6.14",
     "vuex": "^3.6.2"
